backend/requirements.txt
requests
redis
Flask
Flask-Caching
APScheduler
pymysql
mysql-connector-python
flask-cors
